+++ src/blake2s.c
@@ -1,24 +1,97 @@
 #include "blake2s.h"
-#include "utils.h"
 #include <math.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 
+
+/**
+ * Helper function to perform rotation of a 32 bit int
+ */
+uint32_t rotr32(const uint32_t w, const unsigned c)
+{
+  return (w >> c) | (w << (32 - c));
+}
+
+
+/**
+ * Helper function to load 32 bytes at a time into src
+ */
+uint32_t load32( const void *src )
+{
+#if defined(NATIVE_LITTLE_ENDIAN)
+  uint32_t w;
+  memcpy(&w, src, sizeof w);
+  return w;
+#else
+  const uint8_t *p = ( const uint8_t * )src;
+  return (( uint32_t )( p[0] ) <<  0) |
+         (( uint32_t )( p[1] ) <<  8) |
+         (( uint32_t )( p[2] ) << 16) |
+         (( uint32_t )( p[3] ) << 24) ;
+#endif
+}
+
+/**
+ * Stores 16 bit w into a dst
+ */
+void store16( void *dst, uint16_t w )
+{
+#if defined(NATIVE_LITTLE_ENDIAN)
+  memcpy(dst, &w, sizeof w);
+#else
+  uint8_t *p = ( uint8_t * )dst;
+  *p++ = ( uint8_t )w; w >>= 8;
+  *p++ = ( uint8_t )w;
+#endif
+}
+
+/**
+ * Stores w into a dst
+ */
+void store32(void* dst, uint32_t w)
+{
+#if defined(NATIVE_LITTLE_ENDIAN)
+  memcpy(dst, &w, sizeof w);
+#else
+  uint8_t* p = (uint8_t*)dst;
+  p[0] = (uint8_t)(w >> 0);
+  p[1] = (uint8_t)(w >> 8);
+  p[2] = (uint8_t)(w >> 16);
+  p[3] = (uint8_t)(w >> 24);
+#endif
+}
+
+
+/**
+ * Blake2s increment counter
+ */
+void blake2s_increment_counter(blake2s_state* S, const uint32_t inc)
+{
+  S->t[0] += inc ;
+  S->t[1] += (S->t[0] < inc);
+}
+
+/**
+ * Mix function
+ */
 static void G(uint32_t v[16], int a, int b,int c,int d, int32_t x, int32_t y) {                      
  
-    v[a] = v[a] + v[b] + x;
-    v[d] = rotr32(v[d] ^ v[a] , 16);
-    v[c] = v[c] + v[d];
-    v[b] = rotr32(v[b] ^ v[c] , 12);
+  v[a] = v[a] + v[b] + x;
+  v[d] = rotr32(v[d] ^ v[a] , 16);
+  v[c] = v[c] + v[d];
+  v[b] = rotr32(v[b] ^ v[c] , 12);
 
-    v[a] = v[a] + v[b] + y;
-    v[d] = rotr32(v[d] ^ v[a] , 8);
-    v[c] = v[c] + v[d];
-    v[b] = rotr32(v[b] ^ v[c] , 7);         
+  v[a] = v[a] + v[b] + y;
+  v[d] = rotr32(v[d] ^ v[a] , 8);
+  v[c] = v[c] + v[d];
+  v[b] = rotr32(v[b] ^ v[c] , 7);         
 
 }
 
+/**
+ * Compress function
+ */
 static void F(blake2s_state* S, uint8_t block[BLAKE2S_BLOCKBYTES])
 {
   size_t i, j;
@@ -57,6 +130,9 @@ static void F(blake2s_state* S, uint8_t block[BLAKE2S_BLOCKBYTES])
   }
 }
 
+/**
+ * State initialization function
+ */
 int blake2s_init(blake2s_state* S, size_t outlen, const void* key, size_t keylen)
 {
   /*initialize key*/
@@ -93,7 +169,9 @@ int blake2s_init(blake2s_state* S, size_t outlen, const void* key, size_t keylen
   return 0;
 }
 
-
+/**
+ * State update function
+ */
 int blake2s_update(blake2s_state* S, const void* input_buffer, size_t inlen)
 {
   unsigned char* in = (unsigned char*)input_buffer;
@@ -109,6 +187,9 @@ int blake2s_update(blake2s_state* S, const void* input_buffer, size_t inlen)
   return 0;
 }
 
+/**
+ * Final function
+ */
 int blake2s_final(blake2s_state* S, void* out, size_t outlen)
 {
   uint8_t buffer[BLAKE2S_OUTBYTES] = { 0 };
@@ -123,7 +204,9 @@ int blake2s_final(blake2s_state* S, void* out, size_t outlen)
   return 0;
 }
 
-
+/**
+ * Main function
+ */
 int blake2s(void* output, size_t outlen, const void* input, size_t inlen,
         const void* key, size_t keylen)
 {
