+++ src/test.c
@@ -8,7 +8,6 @@
 */
 
 #include "blake2s.h"
-#include "utils.h"
 #include <stdio.h>
 #include <string.h>
